Using oop principles
