+++ src/main/java/Tasklist.java
@@ -1,0 +1,17 @@
+import java.util.ArrayList;
+
+public class Tasklist {
+    public static ArrayList<Task> dukeList;
+
+    public Tasklist() {
+        dukeList = new ArrayList<>();
+    }
+
+    public static void add(Task task) {
+        dukeList.add(task);
+    }
+
+    public static void delete(int taskIndex) {
+        dukeList.remove(taskIndex);
+    }
+}
